feat(Products): add category product route
The new route will reuse the same Product pages taking category parameter from the router params
