+++ frontend/src/pages/index.ts
@@ -1,2 +1,1 @@
-export * from './AllProducts';
-export * from './CategoryProducts';
+export * from './Products';
